Add pair trade subroutine
Adds !pairs command that allows the user to see pair trade statistics for two tickers.

There's a scanner included. It works but currently takes ~15 minutes to scan the NASDAQ, so I commented out the function call in the main app script.

Can be run locally but we can hold off on pushing Discord functionality until a more efficient scan can be developed.